prompt before deletion

+++ packages/frontend/src/CourseView.tsx
@@ -7,6 +7,7 @@ import { Course } from "./types";
 import { useLoaderData } from "react-router-dom";
 import { Input } from "./components/ui/input";
 import { Label } from "./components/ui/label";
+import { Edit, Trash } from "lucide-react";
 
 const CourseDetailEditForm = ({
   course,
@@ -114,6 +115,9 @@ export const CourseView = () => {
   const [isEditMode, setIsEditMode] = useState(false);
 
   const deleteCourse = async () => {
+   const shouldDelete = window.confirm('Are you sure you want to delete this course?');
+   if (!shouldDelete) return;
+
     try {
       await fetch(`${BASE_URL}/courses/${course.id}`, {
         method: "DELETE",
@@ -125,8 +129,7 @@ export const CourseView = () => {
 
   return (
     <Card
-      className="max-w-xs rounded overflow-ellipsis shadow-lg min-h-[175px] flex flex-col
-cursor-default hover:cursor-pointer
+      className="p-4 flex flex-col rounded overflow-ellipsis shadow-lg min-h-[175px]cursor-default hover:cursor-pointer
       "
     >
       <CardHeader className="h-2/3">
@@ -149,10 +152,14 @@ cursor-default hover:cursor-pointer
               Status: <Badge>{status}</Badge>
             </p>
           </CardContent>
-          <Button onClick={deleteCourse}>Delete course</Button>
-          <Button onClick={() => setIsEditMode(true)}>
-            Edit course details
-          </Button>
+          <div className="flex gap-4 self-center">
+            <Button onClick={deleteCourse} variant="destructive">
+              <Trash />
+            </Button>
+            <Button onClick={() => setIsEditMode(true)}>
+              <Edit />
+            </Button>
+          </div>
         </>
       )}
     </Card>
